use better types. Figure out how to add signed to unsigned values.

+++ src/main.zig
@@ -111,8 +111,8 @@ pub const OrdinalDirections = [_]Direction{
 pub const Ivec3 = struct { x: i32, y: i32, z: i32 };
 pub const Ivec2 = struct { x: i32, y: i32 };
 
-pub const Uvec3 = struct { x: u16, y: u16, z: u16 };
-pub const Uvec2 = struct { x: u16, y: u16 };
+pub const Uvec3 = struct { x: usize, y: usize, z: usize };
+pub const Uvec2 = struct { x: usize, y: usize };
 
 pub const Vec3 = struct { x: f32, y: f32, z: f32 };
 pub const Vec2 = struct { x: f32, y: f32 };
